feat(All): Change Country keyword to visitedCountry
