+++ src/lib/persons.ts
@@ -1,8 +1,8 @@
-import { getAllItems, personsPageItemMapper } from './get-all-items'
 import transformJsonSchema, { errorSchema, type ErrorOptionalMessage } from '../models/person'
 import type { Gallery, Person } from '../types/common'
 import type { All } from '../types/pages'
-import { handleLibraryError } from './utils'
+import { handleLibraryError } from './errors'
+import { getAllItems, personsPageItemMapper } from './get-all-items'
 import { readPersons } from './xml'
 
 type Envelope = { body: Person[], status: number }
