+++ src/lib/utils.ts
@@ -1,14 +1,13 @@
 import { glob as globNpm } from 'glob'
 import mime from 'mime-types'
 import path from 'node:path'
-import { z } from 'zod/v4-mini'
-
 import configFile from '../models/config'
 import type {
   Gallery,
   Item,
   XmlItem,
 } from '../types/common'
+import { handleLibraryError, isStandardError, isValidStringArray, isZodError, simplifyZodMessages } from './errors'
 import { type Filesystem } from './filesystems'
 
 const type = (filepath: string): string => {
@@ -76,49 +75,6 @@ function customMime(rawExtension: string) {
   }
 
   return null
-}
-
-function isStandardError(error: unknown): error is Error {
-  if (error instanceof Error) return true
-  if ('message' in (error as any) && 'stack' in (error as any)) {
-    return true
-  }
-  return false
-}
-
-function isZodError(error: unknown): error is z.core.$ZodError {
-  if (error instanceof z.core.$ZodError) return true
-  return false
-}
-
-function simplifyZodMessages(error: z.core.$ZodError) {
-  return error?.issues.reduce((prev: string, curr: z.core.$ZodIssue) => {
-
-    if (prev === '') prev += curr.message
-
-    else prev += `; ${curr.message}`
-    return prev
-  }, '')
-}
-
-function handleLibraryError(err: unknown, message: string, returnEnvelope: boolean, errorSchema: any) {
-  if (returnEnvelope) {
-    if (isStandardError(err)) {
-      return { body: errorSchema(err.message), status: 500 }
-    }
-    return { body: errorSchema(message), status: 404 }
-  }
-
-  console.error('ERROR', message, err)
-  throw err
-}
-
-function isValidStringArray<T extends string = string>(arr: unknown): arr is T[] {
-  return (
-    Array.isArray(arr) &&
-    arr.length > 0 &&
-    arr.every((item): item is T => typeof item === 'string' && item.trim().length > 0)
-  )
 }
 
 function utils() {
@@ -209,5 +165,5 @@ function utils() {
 }
 
 export default utils
-export { isStandardError, isValidStringArray, isZodError, simplifyZodMessages, handleLibraryError }
+export { handleLibraryError, isStandardError, isValidStringArray, isZodError, simplifyZodMessages }
